google-tasks-to-markdown/googleutils.py
from googleapiclient.discovery import build, Resource
import httplib2
from oauth2client import tools
from oauth2client.client import AccessTokenRefreshError
from oauth2client.client import OAuth2WebServerFlow
from oauth2client.file import Storage
import os


# Where OAuth2 creds get saved locally on disk.
CREDENTIALS_PATH = 'credentials.dat'

# The scope URL for read/write access to a user's calendar data
GOOGLE_API_SCOPE = 'https://www.googleapis.com/auth/tasks.readonly'


def build_google_tasks_resource() -> Resource:
    http = get_google_http_client()

    # The apiclient.discovery.build() function returns an instance of an API service
    # object can be used to make API calls. The object is constructed with
    # methods specific to the calendar API. The arguments provided are:
    #   name of the API ('calendar')
    #   version of the API you are using ('v3')
    #   authorized httplib2.Http() object that can be used for API calls
    return build('tasks', 'v1', http=http)


def get_google_http_client() -> Resource:
    # Create a flow object. This object holds the client_id, client_secret, and
    # scope. It assists with OAuth 2.0 steps to get user authorization and
    # credentials.
    flow = OAuth2WebServerFlow(
        os.getenv('GOOGLE_CLIENT_ID'),
        os.getenv('GOOGLE_CLIENT_SECRET'),
        GOOGLE_API_SCOPE,
    )

    # Create a Storage object. This object holds the credentials that your
    # application needs to authorize access to the user's data. The name of the
    # credentials file is provided. If the file does not exist, it is
    # created. This object can only hold credentials for a single user, so
    # as-written, this script can only handle a single user.
    storage = Storage(CREDENTIALS_PATH)

    # The get() function returns the credentials for the Storage object. If no
    # credentials were found, None is returned.
    credentials = storage.get()

    # If no credentials are found or the credentials are invalid due to
    # expiration, new credentials need to be obtained from the authorization
    # server. The oauth2client.tools.run_flow() function attempts to open an
    # authorization server page in your default web browser. The server
    # asks the user to grant your application access to the user's data.
    # If the user grants access, the run_flow() function returns new credentials.
    # The new credentials are also stored in the supplied Storage object,
    # which updates the credentials.dat file.
    if credentials is None or credentials.invalid:
        credentials = tools.run_flow(flow, storage, tools.argparser.parse_args())

    # Create an httplib2.Http object to handle our HTTP requests, and authorize it
    # using the credentials.authorize() function.
    http = httplib2.Http()
    http = credentials.authorize(http)

    return http
