google-tasks-to-markdown/requirements.txt
google-api-python-client==1.7.7
oauth2client==4.1.3
python-dotenv==0.10.1
